Fix selection when multiple tags exist
Now instead of calling :stjump twice, we craft a regexp that matches
either a name tag or a file tag.

We also check whether the tag exists prior to calling stjump, although
that's not really necessary.  It opens the way to avoiding :stjump
altogether in the future, if len(found) == 1.