Hack to deal with parenthesized imports
We'll now recognize and support imports of the form

    from pkg import (
        name1,
        name2,
        name3,
    )

There's also a bugfix where in some circumstances we'd be left with a
search highlighting every beginning of a line.

There's also :ImportName! and :ImportNameHere! that add the import but
maintain the cursor position.

There's also some optional integration with the ALE plugin, except it
doesn't seem to be very effective :/