Speed up the case when there's only one matching tag
Avoid opening a new split window and loading a file when there's only
one matching tag, instead take the tag's filename from the taglist().
This also avoids problems with swap files in case that file is open in a
different Vim instance etc.

When there are several matching tags we still do the :tjump so the user
gets a selection menu.