Bugfix: do not insert first import after actual code
In a new file with some code but no imports, skipping the first
paragraph would insert the import _after_ the existing code, which is
not what the user wants.

I've wanted to fix this bug for so long, but I was afraid of breaking
something else without a test suite.  Now I have a test suite.  And I
was right: I would've broken a lot of things without any tests.

(It's possible that I still have broken things.)