Fix incorrect skipping of the module docstring
The problem: when the first line of a script is """, then

     0/^"""/;/^"""/

finds the closing `"""` for the `0/^"""/` bit, and then wraps around and
finds the opening `"""` for the `;/^"""/` bit.

Unfortunately I was unable to reproduce the problem in Vader tests.  I
don't understand why.

I thought the initial 0 was supposed to make 0/^"""/ accept the current
line?  No idea why it doesn't work in real life.  No idea why it works
in Vader tests.  Some setting from my .vimrc?  No time to dig deeper.