Handle module imports correctly
When the tag you've selected is a file tag (i.e. module.py), we need to
use

    from package import module

instead of

    from package.module import module

This is easy to check when we're using taglist() and have the tag dict
with the 'kind' set to 'F'.  It's a bit harder when we use :stjump and
only know the filename.

I tried to make it somewhat smarter so that if you have a module
thing.py with a function thing() inside it, you can still get

    from package.thing import thing

Unfortunately this will happen even if you select thing.py in the
:stjump menu.

A smarter fix using gettagstack() should be possible.

Closes #14.